Fix Windows Plugin ERRORLEVEL 0 so that %ERRORLEVEL% == "0" Fix Windows Batch script write in newline('\n' -> '\r\n') Fix Result Excel File Name Type Fix Plugin Variable Name

+++ plugins/WINDOWS/OS/W-01.py
@@ -9,36 +9,32 @@ class windowsosw01(Plugin):
 echo     ^<infoElement code="%CODE001%"^> >> %RESULT_COLLECT_FILE%
 
 net user Administrator | findstr /bic:"Account active" > administrator_tmp.txt
-if ERRORLEVEL 0 (
-	call :base64encode administrator_tmp.txt
-	echo         ^<command name="ADMIN_ACCOUNT"^>^<!^[CDATA^[ >> %RESULT_COLLECT_FILE%
-	for /f "delims=" %%a in (base64.txt) do echo %%a >> %RESULT_COLLECT_FILE%
-	echo         ^]^]^>^</command^> >> %RESULT_COLLECT_FILE%
+if "%ERRORLEVEL%" == "0" (
+    call :base64encode administrator_tmp.txt
+    echo         ^<command name="ADMIN_ACCOUNT"^>^<!^[CDATA^[ >> %RESULT_COLLECT_FILE%
+    for /f "delims=" %%a in (base64.txt) do echo %%a >> %RESULT_COLLECT_FILE%
+    echo         ^]^]^>^</command^> >> %RESULT_COLLECT_FILE%
 )
-if exist administrator_tmp.txt (
-	del /q administrator_tmp.txt
-)
+if exist administrator_tmp.txt del /q administrator_tmp.txt
 
 echo     ^</infoElement^> >> %RESULT_COLLECT_FILE%
 
 secedit /export /cfg secpolicy_tmp.txt > nul
 type secpolicy_tmp.txt | more > secpolicy.txt
 del /q secpolicy_tmp.txt
-if ERRORLEVEL 0 (
-	call :fileCheckSum secpolicy.txt, checksumvalue
-	if not "%checksumvalue%" == "DUP" (
-		echo         ^<fileInfo^> >> %RESULT_FILE_DATA_FILE%
-		echo             ^<filePath checksum="%checksumvalue%"^>^<!^[CDATA^[Local Security Policy^]^]^>^</filePath^> >> %RESULT_FILE_DATA_FILE%
-		call :base64encode secpolicy.txt
-		echo             ^<fileData^>^<!^[CDATA^[ >> %RESULT_FILE_DATA_FILE%
-		for /f "delims=" %%a in (base64.txt) do echo %%a >> %RESULT_FILE_DATA_FILE%
-		echo             ^]^]^>^</fileData^> >> %RESULT_FILE_DATA_FILE%
-		echo         ^</fileInfo^> >> %RESULT_FILE_DATA_FILE%  
-	)
+if "%ERRORLEVEL%" == "0" (
+call :fileCheckSum secpolicy.txt, checksumvalue
+    if not "%checksumvalue%" == "DUP" (
+        echo         ^<fileInfo^> >> %RESULT_FILE_DATA_FILE%
+        echo             ^<filePath checksum="%checksumvalue%"^>^<!^[CDATA^[Local Security Policy^]^]^>^</filePath^> >> %RESULT_FILE_DATA_FILE%
+        call :base64encode secpolicy.txt
+        echo             ^<fileData^>^<!^[CDATA^[ >> %RESULT_FILE_DATA_FILE%
+        for /f "delims=" %%a in (base64.txt) do echo %%a >> %RESULT_FILE_DATA_FILE%
+        echo             ^]^]^>^</fileData^> >> %RESULT_FILE_DATA_FILE%
+        echo         ^</fileInfo^> >> %RESULT_FILE_DATA_FILE%
+    )
 )
-if exist secpolicy.txt (
-	del /q secpolicy.txt
-)
+if exist secpolicy.txt del /q secpolicy.txt
 
 echo %CODE001% Collect
 		"""
