+++ plugins/WINDOWS/OS/W-35.py
@@ -9,19 +9,17 @@ class windowsosw35(Plugin):
 echo     ^<infoElement code="%CODE035%"^> >> %RESULT_COLLECT_FILE%
 
 reg query "HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\RemoteRegistry" /s /v Start > remote_reg_tmp.txt
-if ERRORLEVEL 0 (
+if "%ERRORLEVEL%" == "0" (
     call :base64encode remote_reg_tmp.txt
     echo         ^<command name="REMOTE_REGISTRY_REG"^>^<!^[CDATA^[ >> %RESULT_COLLECT_FILE%
     for /f "delims=" %%a in (base64.txt) do echo %%a >> %RESULT_COLLECT_FILE%
     echo         ^]^]^>^</command^> >> %RESULT_COLLECT_FILE%
 )
-if exist remote_reg_tmp.txt (
-    del /q remote_reg_tmp.txt
-)
+if exist remote_reg_tmp.txt del /q remote_reg_tmp.txt
 
 echo     ^</infoElement^> >> %RESULT_COLLECT_FILE%
 
-echo %CODE035% Collect		
+echo %CODE035% Collect	
 		"""
 		self.codeExcute = "set CODE035=W-35"
 		self.description = {
@@ -36,7 +34,7 @@ echo %CODE035% Collect
 		self.fullString = [self.code, '양호', self.stat, self.description]
 
 	def analysisFunc(self, sysInfo, infoDict, fileDict):
-		vulCnt = 0
+		vulCnt = 2
 		chkNum = 0
 		desStr = """
 [참고]
@@ -49,11 +47,12 @@ echo %CODE035% Collect
 		if chkNum > 0:
 			serviceStr = self.stat['Remote Registry_SYS'].replace('\n', '')
 			self.stat.update({'Remote Registry_SYS': f'{serviceStr} Service Running(!)\n'})
-			vulCnt += 1
+		else:
+			vulCnt -= 1
 
 		if infoDict:
 			if 'REMOTE_REGISTRY_REG' in infoDict[0]:
-				vulCnt += self.cmdStrGetValue('CMD:Remote Registry 서비스 시작 유형', infoDict[0], 'REMOTE_REGISTRY_REG',
+				vulCnt -= self.cmdStrGetValue('CMD:Remote Registry 서비스 시작 유형', infoDict[0], 'REMOTE_REGISTRY_REG',
 				                              '^[ \t]*Start.*0x(.)$', '2', '==')
 				valueStr = self.stat['CMD:Remote Registry 서비스 시작 유형']
 				self.stat.update({'CMD:Remote Registry 서비스 시작 유형': f'{valueStr} {desStr}\n'})
